01/requirements.txt
numpy
scipy
opencv-python-headless
networkx
Pillow
tqdm
